feat: automatically install Sprocket from GitHub releases.
This commit adds a feature to the VSCode extension to automatically download
the latest stable Sprocket from GitHub releases.

It does this by first querying `crates.io` for the latest stable `sprocket`
crate version; it then fetches the GitHub release artifact corresponding to
that version.

The extension will also check for newer versions to what has been previously
downloaded and prompt to download and update to the next version.

A new setting has been added to control checking for updates on startup.